requirements.txt
fastapi
uvicorn
boto3
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
